+++ R/greprows.R
@@ -4,6 +4,11 @@
 #' `greprows()` searches for pattern matches within a data frames columns and
 #' returns the related rows or row indices.
 #'
+#' `grepvrows()` is identical to `greprows()` except with the default
+#' `value = TRUE`.
+#'
+#' `greplrows()` returns a logical vector (match or not for each row of dat).
+#'
 # -------------------------------------------------------------------------
 #' @param dat Data frame
 #'
@@ -44,6 +49,11 @@
 #' @seealso [grep()]
 #'
 # -------------------------------------------------------------------------
+#' @export
+#' @name greprows
+NULL
+
+#' @rdname greprows
 #' @export
 greprows <- function(
     dat,
@@ -97,4 +107,58 @@ greprows <- function(
 
     # return the values or the index
     if (value) dat[idx,,drop = FALSE] else idx
+}
+
+
+#' @rdname greprows
+#' @export
+greplrows <- function(
+    dat,
+    pattern,
+    cols = NULL,
+    ignore.case = FALSE,
+    perl = FALSE,
+    fixed = FALSE,
+    invert = FALSE
+) {
+
+    index <- greprows(
+        dat = dat,
+        pattern = pattern,
+        cols = cols,
+        value = FALSE,
+        ignore.case = ignore.case,
+        perl = perl,
+        fixed = fixed,
+        invert = invert
+    )
+
+    out <- logical(length = nrow(dat))
+    out[index] <- TRUE
+    out
+}
+
+
+#' @rdname greprows
+#' @export
+grepvrows <- function(
+        dat,
+        pattern,
+        cols = NULL,
+        value = TRUE,
+        ignore.case = FALSE,
+        perl = FALSE,
+        fixed = FALSE,
+        invert = FALSE
+) {
+    greprows(
+        dat = dat,
+        pattern = pattern,
+        cols = cols,
+        value = value,
+        ignore.case = ignore.case,
+        perl = perl,
+        fixed = fixed,
+        invert = invert
+    )
 }